fix(parser): get tests running again
