+++ package-parser/package_parser/processing/api/documentation_parsing/_DefaultDocumentationParser.py
@@ -1,16 +1,12 @@
-import re
-from typing import Optional, Tuple
-
 import astroid
-import numpydoc.docscrape
-from numpydoc.docscrape import NumpyDocString
 
-from ._AbstractDocumentationParser import AbstractDocumentationParser
-from ._APIElementDocumentation import (
+from package_parser.processing.api.model import (
     ClassDocumentation,
     FunctionDocumentation,
     ParameterDocumentation,
+    ParameterAssignment
 )
+from ._AbstractDocumentationParser import AbstractDocumentationParser
 from ._get_full_docstring import get_full_docstring
 
 
@@ -34,6 +30,9 @@ class DefaultDocumentationParser(AbstractDocumentationParser):
         )
 
     def get_parameter_documentation(
-        self, function_node: astroid.FunctionDef, parameter_name: str
+        self,
+        function_node: astroid.FunctionDef,
+        parameter_name: str,
+        parameter_assigned_by: ParameterAssignment,
     ) -> ParameterDocumentation:
         return ParameterDocumentation()
