+++ package-parser/package_parser/processing/api/documentation_parsing/__init__.py
@@ -1,9 +1,4 @@
 from ._AbstractDocumentationParser import AbstractDocumentationParser
-from ._APIElementDocumentation import (
-    ClassDocumentation,
-    FunctionDocumentation,
-    ParameterDocumentation,
-)
 from ._DefaultDocumentationParser import DefaultDocumentationParser
 from ._get_full_docstring import get_full_docstring
 from ._NumpyDocParser import NumpyDocParser
