+++ package-parser/tests/processing/api/documentation/test_APIElementDocumentation.py
@@ -1,5 +1,6 @@
 import pytest
-from package_parser.processing.api.documentation import (
+
+from package_parser.processing.api.model import (
     ClassDocumentation,
     FunctionDocumentation,
     ParameterDocumentation,
