+++ package-parser/tests/processing/dependencies/test_get_dependency.py
@@ -1,6 +1,6 @@
 import spacy
-from package_parser.processing.api.documentation import ParameterDocumentation
-from package_parser.processing.api.model import Parameter, ParameterAssignment
+
+from package_parser.processing.api.model import ParameterDocumentation, Parameter, ParameterAssignment
 from package_parser.processing.dependencies import (
     Action,
     Condition,
